Put FromField for TimeUnitField in timeunit mod
